Safari-611.1.10/LayoutTests/fast/url/data-url-source-expected.txt

PASS internals.xhrResponseSource(xhr) is "Network"
PASS successfullyParsed is true

TEST COMPLETE
